% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_bivariate.R
\name{calc_kappa_lb_2d}
\alias{calc_kappa_lb_2d}
\title{Calculate the lower bound of the condition number for the covariance matrix
of two random variables with a given correlation}
\usage{
calc_kappa_lb_2d(corr)
}
\arguments{
\item{corr}{The correlation coefficient, a numeric value between -1 and 1.}
}
\value{
The numeric lower bound of the spectral condition number.
}
\description{
The condition number of a covariance matrix \eqn{\Sigma = \text{Cov}(W)} for
a bivariate random vector \eqn{W = (W_1, W_2)} is bounded below by
\deqn{\kappa(\Sigma) \geq \frac{1 + |\rho|}{1 - |\rho|}}
where \eqn{\rho = \text{Corr}(W_1,W_2) \in (-1, 1)} is the underlying correlation
between the two component variables.
}
