% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_OLS.R
\name{compute_split_OLS}
\alias{compute_split_OLS}
\title{Compute OLS solutions for a split}
\usage{
compute_split_OLS(C1, C2, w.mat, y.vec, labels = NULL, ...)
}
\arguments{
\item{C1}{A vector of indices (logical or integer) for the first child node. See \code{\link[=solve_OLS]{solve_OLS()}}.}

\item{C2}{A vector of indices (logical or integer) for the second child node. See \code{\link[=solve_OLS]{solve_OLS()}}.}

\item{w.mat}{The full matrix of regressors.}

\item{y.vec}{The full vector of outcomes.}

\item{labels}{An optional character vector to name the resulting coefficients.}

\item{...}{Additional arguments passed to \code{\link[=solve_OLS]{solve_OLS()}}.}
}
\value{
A list containing the named coefficient vectors for C1 and C2.
}
\description{
A wrapper function that computes the OLS solutions for the two child nodes
(C1 and C2) induced by a split.
}
\keyword{internal}
