% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_OLS.R
\name{compute_theta_hat}
\alias{compute_theta_hat}
\title{Compute empirical solution for the local parameter estimates under a varying-coefficient model
fit across a candidate split}
\usage{
compute_theta_hat(C1, C2, w.mat, y.vec, labels = NULL, ...)
}
\arguments{
\item{C1}{A vector of indices (logical or integer) for the first child node. See \code{\link[=solve_OLS]{solve_OLS()}}.}

\item{C2}{A vector of indices (logical or integer) for the second child node. See \code{\link[=solve_OLS]{solve_OLS()}}.}

\item{w.mat}{The full matrix of regressors.}

\item{y.vec}{The full vector of outcomes.}

\item{labels}{An optional character vector to name the resulting coefficients.}

\item{...}{Additional arguments passed to \code{\link[=solve_OLS]{solve_OLS()}}.}
}
\value{
A list containing the named coefficient vectors for C1 and C2.
}
\description{
A wrapper function that computes child solutions \eqn{\hat\theta_{C_1}} and \eqn{\hat\theta_{C_2}}
for the optimal local parameters fit over the data in \eqn{C_1} and \eqn{C_2}, respectively.
}
\keyword{internal}
