% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteval-package.R
\docType{package}
\name{criteval-package}
\alias{criteval}
\alias{criteval-package}
\title{criteval: Simulations and Comparisons of Splitting Criteria for Generalized Random Forests}
\description{
An evaluation framework for the simulation and comparison of splitting criteria used by GRF in its recursive partitioning mechanism, particularly for causal forests and multi-dimensional varying-coefficient models.
}
\author{
\strong{Maintainer}: David Fleischer \email{david.p.fleischer@gmail.com}

}
\keyword{internal}
