% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cov.R
\name{generate_AR1}
\alias{generate_AR1}
\title{Generate an AR(1) correlation matrix}
\usage{
generate_AR1(dim, rho)
}
\arguments{
\item{dim}{The dimension of the square matrix.}

\item{rho}{The AR(1) correlation parameter, a value in (-1, 1).}
}
\value{
A \code{dim}-by-\code{dim} matrix with the AR(1) correlation matrix structure.
}
\description{
Creates a square correlation matrix with an Autoregressive(1) structure.
}
