% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_bivariate.R
\name{generate_bivariate}
\alias{generate_bivariate}
\title{Generate bivariate Gaussian samples}
\usage{
generate_bivariate(
  n,
  corr,
  kappa.cov = NULL,
  mu = c(0, 0),
  .tol = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{n}{The number of observations to generate.}

\item{corr}{The desired correlation between the two variables, a value in (-1, 1).}

\item{kappa.cov}{The desired spectral condition number of the covariance matrix. If \code{NULL},
it is set to the minimum possible value for the given \code{corr}. See \code{\link[=calc_kappa_lb_2d]{calc_kappa_lb_2d()}}.}

\item{mu}{A numeric vector of length 2 for the mean of the variables. Defaults to \code{mu = c(0, 0)}.}

\item{.tol}{A small tolerance for numerical comparisons.}
}
\value{
An n-by-2 matrix of regressors. The matrix has attributes
\code{mu}, \code{Sigma}, \code{corr}, and \code{kappa.cov} attached.
}
\description{
Generates a 2-column matrix containing samples of a bivariate Gaussian vector
with a specified underlying correlation coefficient and a specified spectral
condition for its covariance matrix.
}
