% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_cov.R
\name{generate_corr}
\alias{generate_corr}
\title{Generate a correlation matrix}
\usage{
generate_corr(rho = NULL, K = NULL, R = NULL)
}
\arguments{
\item{rho}{The AR(1) correlation parameter, a value in (-1, 1). See \code{\link[=generate_AR1]{generate_AR1()}}.
Only used if \code{R} is \code{NULL}.}

\item{K}{The target dimension for the matrix. Required if \code{R} is \code{NULL}, ignored otherwise.}

\item{R}{An optional, pre-specified correlation matrix. If provided, \code{rho} and \code{K} are ignored.}
}
\value{
A \eqn{K}-by-\eqn{K} correlation matrix.
}
\description{
A helper function to generate a \eqn{K}-by-\eqn{K} correlation matrix, defaulting to an
AR(1) structure if a specific matrix \code{R} is not provided.
}
