% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_integrator.R
\name{precompute_integrals_along_dim}
\alias{precompute_integrals_along_dim}
\title{Pre-computes inner integrals to create a lookup table}
\usage{
precompute_integrals_along_dim(ell, p, theta_FUN_list, grid.size = 200, ...)
}
\arguments{
\item{ell}{The candidate splitting dimension along which we wish to pre-compute values of the
inner integral function. See the \strong{Details} section below.}

\item{p}{The total dimensionality of the variable of integration \eqn{x}.}

\item{theta_FUN_list}{The list of the \eqn{\theta^*(x) = (\theta_1^*(x),\ldots,\theta_K^*(x))}
component functions \eqn{\theta_k^*:[0,1]^p\to\mathbb R}.}

\item{grid.size}{The number of points for the lookup table grid.}

\item{...}{Additional arguments for the integrator passed to \code{\link[cubature:hcubature]{cubature::hcubature()}}.}
}
\value{
An object of class \code{precomputed_integral} for fast querying.
}
\description{
This function performs expensive, high-dimensional integration upfront. For a
given candidate splitting axis/dimension \eqn{\ell \in \{1,\ldots,p\}} over the
covariate space, it integrates the target function \eqn{\theta^*(x)} over the
other \eqn{p - 1} dimensions at various discrete points along the \eqn{\ell}-th axis.
The result is a lookup table that can be queried quickly and used for interpolation
with faster, one-dimensional integration.
}
\details{
Let's say we want to calculate the integral of \eqn{\theta^*(x)} over a hyperrectangle
(child node) \eqn{C = [0,t] \times [0,1]^{p-1}}, where we're assuming for illustration
that the candidate split is along the first dimension \eqn{\ell = 1}. The full integral is
\deqn{\int_{C} \theta^*(x)\,dx = \int_0^t \left( \int_{[0,1]^{p-1}}\theta^*(x_1,\ldots,x_p)\,dx_2\cdots\,dx_p \right)\,dx_1}
This breaks the problem into two parts:
\enumerate{
\item An "inner integral": The part in the parentheses, which integrates over \eqn{p - 1}
dimensions that are not part of the candidate splitting dimension \eqn{\ell = 1}. The result
of this inner integral is a functioin of the one remaining variable, \eqn{x_1}. Call this
resulting function \eqn{h(x_1)}.
\item An "outer integral": A simple 1D integral of \eqn{h(x_1)} with respect to the splitting
variable \eqn{x_1} over the domain of valid splitting thresholds \eqn{t \in [0,1]}.
}
The \code{precompute_integrals_along_dim} function is designed to numerically calculate this
inner integral function, \eqn{h(x_1)}, over a grid of points along the \eqn{x_1}-axis
(or, in general, the inner integral function \eqn{h(x_\ell)} over a grid of points along
the \eqn{x_\ell}-axis).
}
