% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_integrator.R
\name{query_precomputed_integrals}
\alias{query_precomputed_integrals}
\title{Queries the pre-computed lookup table to get parameter function integrals
across a candidate split}
\usage{
query_precomputed_integrals(precomputed_obj, t)
}
\arguments{
\item{precomputed_obj}{The object returned by \code{\link[=precompute_integrals_along_dim]{precompute_integrals_along_dim()}}.}

\item{t}{The splitting threshold \eqn{t \in (0, 1)} along the candidate splitting
axis/dimension corresponding to pre-computed integral object \code{obj}.}
}
\value{
A list containing the integral values for the left child
(\eqn{x_{\ell} \leq t}) and the right child (\eqn{x_{\ell} > t}) that
result from the split along dimension \eqn{\ell \in \{1,\ldots, p\}} at
threshold \eqn{t \in (0,1)}.
}
\description{
Uses fast 1D interpolation and integration on the pre-computed "inner" integral
object return by \code{\link[=precompute_integrals_along_dim]{precompute_integrals_along_dim()}} to find the intergal of
\eqn{\theta^*(x)} over two child regions defined by a binary partition of
the covariate space \eqn{[0,1]^p} into two hyperrectangles. The partition is
defined by splitting parameters \eqn{(\ell, t)}, where \eqn{\ell \in \{1,\ldots, p\}}
denotes the splitting axis/dimension along which the precomputed inner function
values were solved, and splitting threshold \eqn{t\in (0,1)}.
}
\keyword{internal}
