% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_criteval_sim.R
\name{run_criteval_sim}
\alias{run_criteval_sim}
\title{Run the splitting criteria simulation}
\usage{
run_criteval_sim(
  n,
  kappa.ratio,
  rho.W = NULL,
  R.W = NULL,
  q.W,
  mu.W = NULL,
  Sigma.W = NULL,
  p.X,
  sigma.eps,
  theta_FUN_list,
  nu_FUN,
  criteria = criteval::get_criteria(),
  min.node.size = 10,
  .n.skip = 1,
  .grid.size = 200,
  .tol.integrate = 1e-04,
  seed = NULL
)
}
\arguments{
\item{n}{Integer, the total number of observations.}

\item{kappa.ratio}{Numeric, the target condition number ratio for the covariance matrix
of the primary regressors \code{W}. Ignored if a covariance matrix \code{Sigma.W} is provided.}

\item{rho.W}{Numeric, the AR(1) correlation for \code{W} regressors. Ignored if either a
correlation matrix \code{R.W} or a covariance matrix \code{Sigma.W} are provided.}

\item{R.W}{A pre-specified correlation matrix for \code{W}. Ignored if a covariance matrix
\code{Sigma.W} is provided.}

\item{q.W}{Integer, the number of "strong" (high-variance) \code{W} regressors. Ignored if a
covariance matrix \code{Sigma.W} is provided.}

\item{mu.W}{Numeric vector, the mean of \code{W} regressors.}

\item{Sigma.W}{A covariance matrix for \code{W}.}

\item{p.X}{Integer, the dimension of the auxiliary covariates \code{X}.}

\item{sigma.eps}{Numeric, the standard deviation of the error term.}

\item{theta_FUN_list}{A list of functions defining the \code{theta*(x)} components.}

\item{nu_FUN}{A function defining the nuisance component \code{nu(x)}.}

\item{criteria}{A list of \code{criteval_criterion} objects to evaluate. Defaults to all
criteria from \code{criteval::get_criteria()}.}

\item{min.node.size}{Integer, the minimum number of observations in any child node.}

\item{.n.skip}{Integer, interval for skipping split points to speed up computation.
Computes criteria every \code{.n.skip} values. Defaults to 1.}

\item{.grid.size}{Integer, the resolution for the numerical integration lookup table.}

\item{.tol.integrate}{Tolerance parameter passed to \code{\link[cubature:hcubature]{cubature::hcubature()}} when
performing numerical integration.}

\item{seed}{Integer, an optional random seed for reproducibility.}
}
\description{
Orchestrates a simulation to evaluate and compare splitting criteria in their
ability to detect heterogeneity in an unobserved parameter \eqn{\theta^*(x)}.
The task of the criteria is to select a single binary axis-aligned split of
the data such as to separate the covariate space into dissimilar regions
with respect to \eqn{\theta^*(x)}.
}
\note{
The numerical integration for \eqn{theta^*(x)} assumes that the auxiliary
covariates \eqn{X} are uniformly distributed over the covariate space \eqn{[0,1]^p}.
Therefore, the \code{rho.X} parameter for the \code{data_generator()} is intentionally not
exposed because the general Gaussian copula would violate these assumptions and
lead to spurrious expectations \eqn{\mathbb E[\theta^*(X) \mid X \in C_j]}.
}
