% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_bivariate.R
\name{validate_corr_kappa_2d}
\alias{validate_corr_kappa_2d}
\title{Validate the feasibility between a correlation coefficient and a condition number}
\usage{
validate_corr_kappa_2d(
  corr,
  kappa.cov = NULL,
  .tol = sqrt(.Machine$double.eps)
)
}
\arguments{
\item{corr}{The correlation coefficient between two random variables.}

\item{kappa.cov}{The spectral condition number of a 2x2 covariance matrix for random variables
with correlation \code{corr}.}

\item{.tol}{A small tolerance for numerical comparisons.}
}
\value{
Returns a named list containing the original input correlation \code{corr},
the validated condition number \code{kappa.cov}, and the lower bound \code{kappa.lb} for
the condition number of the covariance between two variables with correlation \code{corr}.
See \code{\link[=calc_kappa_lb_2d]{calc_kappa_lb_2d()}}.
}
\description{
Checks whether the desired spectral condition number of some 2x2 covariance matrix is
feasible given the supplied correlation coefficient.
}
