% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/criteria.R
\name{get_criteria}
\alias{get_criteria}
\title{Get a list of predefined criterion objects}
\usage{
get_criteria()
}
\value{
A named list of \code{criteval_criterion} objects.
}
\description{
This function returns a named list of predefined \code{criteval_criterion} objects.
This serves as a central registry for all available criteria, making it easy
to extend the framework by simply adding new definitions here.
}
